+++ src/krylov/krylov_fctry.f90
@@ -43,6 +43,7 @@ submodule (krylov) krylov_fctry
   use fusedcg_device, only : fusedcg_device_t
   use fusedcg_cpld_device, only : fusedcg_cpld_device_t
   use bicgstab, only : bicgstab_t
+  use bicgstab_cpld, only : bicgstab_cpld_t
   use gmres, only : gmres_t
   use cheby, only : cheby_t
   use cheby_device, only : cheby_device_t
@@ -55,7 +56,7 @@ submodule (krylov) krylov_fctry
   implicit none
 
   ! List of all possible types created by the factory routine
-  character(len=20) :: KSP_KNOWN_TYPES(9) = [character(len=20) :: &
+  character(len=20) :: KSP_KNOWN_TYPES(10) = [character(len=20) :: &
        "cg", &
        "pipecg", &
        "fused_cg", &
@@ -63,6 +64,7 @@ submodule (krylov) krylov_fctry
        "gmres", &
        "cheby", &
        "bicgstab", &
+       "coupled_bicgstab", &
        "fused_coupled_cg", &
        "coupled_cg"]
 
@@ -175,6 +177,13 @@ contains
 
     case ('bicgstab')
        allocate(bicgstab_t::object)
+
+    case ('coupled_bicgstab')
+       if (NEKO_BCKND_DEVICE .eq. 1) then
+          call neko_error('Coupled BiCGStab is not supported on devices')
+       else
+          allocate(bicgstab_cpld_t::object)
+       end if
 
     case default
        do i = 1, krylov_registry_size